fix issue with piston pushing invisible blocks potions dont always apply gravity when running on a server. though it still looks like they do.

+++ src/main/java/com/arcanist/magick/entity/FirePearlEntity.java
@@ -45,8 +45,8 @@ public class FirePearlEntity extends ThrownItemEntity {
     protected void onEntityHit(EntityHitResult entityHitResult) { // called on entity hit.
         super.onEntityHit(entityHitResult);
         Entity entity = entityHitResult.getEntity();
-        int i = entity instanceof SnowGolemEntity ? 6 : 0;
-        entity.damage(DamageSource.thrownProjectile(this, user), (float) i + damage);
+        float i = entity instanceof SnowGolemEntity ? 6 : 0;
+        entity.damage(DamageSource.thrownProjectile(this, user), i + damage);
         if (entity instanceof LivingEntity) {
             entity.setOnFireFor(200);
         }
